renamed backend files to proper naming schemes
