+++ User.js
@@ -1,6 +1,6 @@
 let express = require("express");
 let router = express.Router();
-let Users = require("./models/Users");
+let Users = require("./models/user_model");
 
 router.post("/", function(req, res) {
   res.json({ msg: "hello" });
